minimarket/requirements.txt
# Framework principal
Flask==3.0.0

# Extensiones de Flask
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1

# Base de datos y autenticación
pyodbc==5.2.0
bcrypt==4.1.2
SQLAlchemy==2.0.43

# Formularios y validación
WTForms==3.1.1

# Manejo de imágenes
pillow==11.3.0

# HTTP requests y APIs
requests==2.31.0

# Configuración y variables de entorno
python-dotenv==1.0.0

# Zona horaria
pytz==2023.3

# Dependencias automáticas (no instalar manualmente)
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3